requirements.txt
# Core dependencies
gradio==4.44.1
groq==0.4.1
python-dotenv==1.0.1

# Scientific computing and ML
numpy==1.26.4
faiss-cpu==1.7.4
sentence-transformers==2.5.1

# Web server
aiohttp==3.9.3
uvicorn==0.27.1 